change size_t to off_t

+++ system/lib/wasmfs/js_api.cpp
@@ -246,7 +246,7 @@ int _wasmfs_close(int fd) {
   return __wasi_fd_close(fd);
 }
 
-int _wasmfs_mmap(size_t length, int prot, int flags, int fd, size_t offset) {
+int _wasmfs_mmap(size_t length, int prot, int flags, int fd, off_t offset) {
   return __syscall_mmap2(0, length, prot, flags, fd, offset);
 }
